moved MobileDeviceWrapper out from Playground
