+++ docuilib/src/theme/ReactLiveScope/index.js
@@ -22,6 +22,7 @@ import SortableList from 'react-native-ui-lib/sortableList';
 import Switch from 'react-native-ui-lib/switch';
 import TextField from 'react-native-ui-lib/textField';
 import * as Playground from './Playground';
+import {MobileDeviceWrapper} from './MobileDeviceWrapper';
 
 Assets.loadAssetsGroup('icons.demo', {
   // chevronDown: require('../../assets/icons/chevronDown.png').default,
@@ -55,7 +56,8 @@ const ReactLiveScope = {
   ...React,
   /* Docs' utils and components */
   Data,
-  ...Playground,
+  Playground,
+  MobileDeviceWrapper,
   /* UI Lib's components */
   ActionBar,
   Assets,
